Handle linking to series page from other pages using a custom URL
In order to reuse the SortableTable component, which is kind of big and
don't allow much control over the table rows, I give a render property
so that I can customize how a column should look with a function.

In this case, the objetive is to render the first column as link to the
series page, and modify the default URL it fetches from the server.

I reused the buildFilter function to do that, so that it takes an
array of key values pairs like [name, value] and build the url from
there.