Prevent duplicatin filter query in odata request url
When filtering by date, it is not allowed to pass two different filter
queries. In this case, the problem was that I was first filtering by
geographic level and then filtering by date if the user wanted to, which
would insert another filter query:

  "$URL?$filter=NIVNOME eq 'Brasil'&$filter=VALDATA ge 2020-02-03"

This is not allowed.

This is patchy and ugly but I don't think I will need to tweak much more
so that's it.