Add feature for user to give boundaries to maps
Now if the user wants to see some data by municipalities within the
state of Rio de Janeiro only, it is possible.

I changed some "terminology":

`geoDivision` refers to how the map is divided and it is encoded by
numbers from 0 (no divisions) to 5 (municipalities), which must be
passed to parameter `resolucao` in a IBGE's Maps API call.

Given a division, choosing a boundary only makes sense when the chart
type is "map", so I extracted that function to the `api/ibge.js` module.

Given the division, only some boundaries make sense (you can't see
macroregions inside Rio de Janeiro), so the function `getTopRegions`
handles that (given a boundary, it gets its IBGE code, the logic is then
to get boundaries with code numbers below it).

`geoBoundary` is a description of a boundary ("Estados", "Regiões").

Given a boundary and that the chart is a map, we get the its possible
values (for example, get all states). We use IBGE's Locations API for
that.

The user then chooses a value and submits the form. In the select input,
the value attribute has the boundary's id (hence `geoBoundaryId`), which
we're going to need to get the map.