Show NoData component when data is loaded but it's empty
This currently can only happen with COVID 19 related series (which are
all empty for some reason), which caused the application to crash.

It could happen to if the user somehow gave a period with no data (which
is possible, e.g. when the data is quadriennal and the user entered some
period where there is no data avaiable (two years in between etc.)).