Get labels for each map region from IBGE API
Before that, I used the information already available at the response
from the Ipeadata array (i.e., for each feature, I would look at the
series array and get the correspoing TERNOME [territory name]).

I already get the IBGE array for the form, so it's already cached and
it's fast anyway. So I fetch it again here to get the labels.