No need to handle form state manually
Material-UI already does this. I just need to handle the submit event.

Also, do not change values based on search params for select inputs and
date inputs because the actual value (in opposition to the "display"
value) is often not obvious (e.g., SERSTATUS has two options, "Ativa"
and "Inativa", which translates to "A" and "I" when we make the request.
I don't think it's really worth to handle all of these.