Show tooltip when hovering map
* Problem

This is very complicated, surprisingly. With chartjs-chart-geo I didn't
need to to anything but react-simple-maps provides no built-in way to
show a tooltip on hover. An example is provided using react-tooltip.

I tried that at first but it was very slow, because we have to isolate
somehow the process of drawing the map and the process of getting the
tooltip's state.

With the react-tooltip approach, there is the major limitation of having
to give the required `data-title` attribute to the ComposableMap
component, which I also have to use when drawing the map so it wasn't
clear to me how I would do that.

Also, the component Geography did not forward refs and therefore the
tooltip would misbehave.

* Solution

The solution was to use the Material-UI Tooltip component in a wrapper
component that renders the map and the tooltip.

The tooltip state must be set inside the ChoroplethMap component
context, but I also do not want to re-render that component when the
Tooltip state changes.

The solution for that is use `React.memo`. I do not fully understand how
exactly it works and I'll read about it.

Also, the tooltip follows the mouse cursor, so there is a onMouseMove
event which would a very heavy burden on the rendering of the tree so we
must absolutely use `React.memo`.