Fix offsetting date logic
For some datasets, it doesn't work to use the odata parameter $top
(regional and categorical data).

The solution is to offset the max date by a given number of periods N to
get the same effect.

This patch just makes more intuitive that the variable `DEFAULT_OFFSET`
means that we want, by default, observations for the last, say, 5
periods.

If the variable is 0, we will get no observations. If it is -1 the date
will be incremented and the effect will be the same.